+++ package.json
@@ -1,0 +1,9 @@
+{
+  "devDependencies": {
+    "babel-cli": "^6.24.1",
+    "babel-preset-env": "^1.6.0"
+  },
+  "scripts": {
+    "start": "babel-node main.js"
+  }
+}
